Clean code and fix some bugs
